requirements.txt
requests
beautifulsoup4
jinja2 
python-dotenv 